Added angular Frontend
Angular project that provides a responsive and interactive frontend for a web application. Implements user authentication and authorization, dynamic navigation, and various UI components. Built with Angular framework and Bootstrap CSS. Developed using modern web development practices and techniques.

+++ Backend/Backend.Domain/DTOs/CSVDto.cs
@@ -17,8 +17,8 @@ namespace Backend.Domain.DTOs
         public int ReturnStationId { get; set; }
         [Ignore]
         public string ReturnStation { get; set; }
-        public int CoveredDistanceInMeters { get; set; }
-        public int DurationInSeconds { get; set; }
+        public double CoveredDistanceInMeters { get; set; }
+        public double DurationInSeconds { get; set; }
         public CSVDto() { }
     }
     public static class JourneyDtoExtensions
